OASIS UBL 2.0 Update Package Release
The UBL 2.0 Update Package (officially known in OASIS terminology
as UBL 2.0 Errata 01) provides UBL 2.0 users with an upgrade to
the December 2006 UBL 2.0 release.  Specifically, it:

 - Corrects a number of typos and editorial errors in the
   documentary portions of the UBL 2.0 schemas and data models

 - Fixes a broken entry for BOSNIA AND HERZEGOVINA in
   cl/gc/default/CountryIdentificationCode-2.0.gc

 - Upgrades all the code list files from genericode 0.4 format to
   genericode 1.0 format and updates the listURI in the Qualified
   Datatype spreadsheet and schemas accordingly

 - Revises cl/gc/special-purpose/PortCode-2.0.gc to add additional
   columns and breaks out subsets of that code list into smaller
   files for ease of implementation

 - Replaces the original val/defaultCodeList.xsl with a version
   that reflects the revised code lists

A detailed record of all changes to the original December 2006 UBL
2.0 data models and schemas can be found in Section 3 of the PDF
file included in the Package.  None of these changes is considered
substantive in the sense that any of them would require
modifications to existing software.